Change to prepared statements + cleanup
