+++ src/main/java/com/napier/sem/App.java
@@ -2,23 +2,26 @@ package com.napier.sem;
 
 import com.napier.sem.db.DbConnection;
 import com.napier.sem.report.Reports;
+import com.napier.sem.sql.AdditionalReports;
 import com.napier.sem.sql.Sql;
 import com.napier.sem.world.City;
+import com.napier.sem.world.PopulationData;
 
 import java.sql.Connection;
 import java.util.ArrayList;
 
 /**
  * Project Authors: Colin, Cameron, Luke, Del
- * Date last modified: 21/02/2021
+ * Date last modified: 22/02/2021
  * Purpose of program: To run SQL queries to satisfy the given requirements
+ * Last modified by: Cameron
  */
 
 public class App
 {
     public static void main(String[] args)
     {
-        //status update
+        // Status update
         System.out.println("Added Con & tidied up tested full - working");
 
         // Database connection setup
@@ -31,7 +34,7 @@ public class App
 
         // Display report from cities7
         Reports report7 = new Reports();
-        report7.printCities(cities7);
+        //report7.printCities(cities7);
 
         // Return all cities in a continent to cities8 object
         Sql query8 = new Sql();
@@ -39,7 +42,7 @@ public class App
 
         // Display report from cities8
         Reports report8 = new Reports();
-        report8.printCities(cities8);
+        //report8.printCities(cities8);
 
         // Return all cities in a region to cities9 object
         Sql query9 = new Sql();
@@ -47,7 +50,7 @@ public class App
 
         // Display report from cities9
         Reports report9 = new Reports();
-        report9.printCities(cities9);
+        //report9.printCities(cities9);
 
         // Return all cities in a region to cities10 object
         Sql query10 = new Sql();
@@ -55,7 +58,7 @@ public class App
 
         // Display report from cities10
         Reports report10 = new Reports();
-        report10.printCities(cities10);
+        //report10.printCities(cities10);
 
         // Return all cities in a district to cities11 object
         Sql query11 = new Sql();
@@ -63,7 +66,7 @@ public class App
 
         // Display report from cities11
         Reports report11 = new Reports();
-        report11.printCities(cities11);
+        //report11.printCities(cities11);
 
         // Return top N populated cities in the world to cities12 object
         Sql query12 = new Sql();
@@ -71,7 +74,7 @@ public class App
 
         // Display report from cities12
         Reports report12 = new Reports();
-        report12.printCities(cities12);
+        //report12.printCities(cities12);
 
         // Return top N populated cities in a continent to cities13 object
         Sql query13 = new Sql();
@@ -79,7 +82,7 @@ public class App
 
         // Display report from cities13
         Reports report13 = new Reports();
-        report13.printCities(cities13);
+        //report13.printCities(cities13);
 
         // Return top N populated cities in a region to cities14 object
         Sql query14 = new Sql();
@@ -87,7 +90,7 @@ public class App
 
         // Display report from cities14
         Reports report14 = new Reports();
-        report14.printCities(cities14);
+        //report14.printCities(cities14);
 
         // Return top N populated cities in the world to cities15 object
         Sql query15 = new Sql();
@@ -95,7 +98,7 @@ public class App
 
         // Display report from cities15
         Reports report15 = new Reports();
-        report15.printCities(cities15);
+        //report15.printCities(cities15);
 
         // Return top N populated cities in the world to cities16 object
         Sql query16 = new Sql();
@@ -103,22 +106,22 @@ public class App
 
         // Display report from cities16
         Reports report16 = new Reports();
-        report16.printCities(cities16);
+        //report16.printCities(cities16);
 
 
         // Additional Information Reports //
         // Report 1:
         // Get population of the world
-        //Long totalPopulation = AdditionalReports.getWorldPopulation(con);
-        //System.out.println("World Population: " + totalPopulation);
+        // Long totalPopulation = AdditionalReports.getWorldPopulation(con);
+        // System.out.println("World Population: " + totalPopulation);
 
         // Report 2:
         // Get population of a continent
         // Set continent to get population of
         // 'Asia','Europe','North America','Africa','Oceania','Antarctica','South America'
-        //String continent = "Europe";
-        //Long continentPopulation = AdditionalReports.getPopulationOfContinent(con, continent);
-        //System.out.println("Population of " + continent + " is " + continentPopulation);
+        // String continent = "Europe";
+        // Long continentPopulation = AdditionalReports.getPopulationOfContinent(con, continent);
+        // System.out.println("Population of " + continent + " is " + continentPopulation);
 
         // Report 3:
         // Get population of a region
@@ -153,17 +156,17 @@ public class App
         // The population of people, people living in cities, and people not living in cities in each continent.
         // Antarctica is not shown as the population is set to 0 in the database //
         //ArrayList<PopulationData> continentPopulationDataList = AdditionalReports.getPopulationDataContinent(con);
-        //AdditionalReports.printPopulationDataContinent(continentPopulationDataList);
+        //Reports.printPopulationDataContinent(continentPopulationDataList);
 
         // Report 2:
         // The population of people, people living in cities, and people not living in cities in each region.
         //ArrayList<PopulationData> regionPopulationDataList = AdditionalReports.getPopulationDataRegion(con);
-        //AdditionalReports.printPopulationDataRegion(regionPopulationDataList);
+        //Reports.printPopulationDataRegion(regionPopulationDataList);
 
         // Report 3:
         // The population of people, people living in cities, and people not living in cities in each country.
         //ArrayList<PopulationData> countryPopulationDataList = AdditionalReports.getPopulationDataCountry(con);
-        //AdditionalReports.printPopulationDataCountry(countryPopulationDataList);
+        //Reports.printPopulationDataCountry(countryPopulationDataList);
 
         // Disconnect from database
         dbc.disconnect();
