+++ src/main/java/com/napier/sem/world/City.java
@@ -9,10 +9,8 @@ package com.napier.sem.world;
 
 public class City {
 
-public String name;
-public String country;
-public String district;
-public String population;
-
-
+    public String name;
+    public String country;
+    public String district;
+    public String population;
 }
